// pwa/src/app/api/dto/auditoria.dto.ts
export interface Auditoria {
    idauditoria:number;
    CUIT: string;
    EstadoAuditoria: string;
    colorEstado: string;
    Prestador: string;
    ProvinciaPrestador: string;
    UGL: string;
    domicilio: string;
    email: string;
    fechainicio: string;
    idestadoauditoria: number;
    idguia: string;
    idprovincia: string;
    idugl: string;
    localidad: string;
    telefono: string;
    versionguia: string;
}